Add files via upload
Created a webpage that displays all the content about nutritious food alongside it includes portals that describe how to maintain a balanced diet and it's consumption on a day-to-day basis. Engaged with Cascading Style Sheets (CSS) to make the page look visually appealing! As you can view with the above files.